[CL-3033] Move modal to its own component
